Risky proposals query
- reusable function for preparing new data in query
