+++ src/pages/FundProposalsRisky/Positions.tsx
@@ -5,10 +5,10 @@ import { disableBodyScroll, clearAllBodyScrollLocks } from "body-scroll-lock"
 
 import { useActiveWeb3React } from "hooks"
 import { RiskyPositionsQuery } from "queries"
+import { prepareRiskyPositions } from "utils"
 import { usePoolContract } from "hooks/usePool"
 import useQueryPagination from "hooks/useQueryPagination"
 import { usePoolMetadata } from "state/ipfsMetadata/hooks"
-import { IRiskyPositionCard } from "interfaces/thegraphs/basic-pools"
 
 import LoadMore from "components/LoadMore"
 import RiskyPositionCard from "components/cards/position/Risky"
@@ -41,28 +41,10 @@ const FundPositionsRisky: FC<IProps> = ({ poolAddress, closed }) => {
     [closed, poolAddress]
   )
 
-  const prepareNewData = (d): IRiskyPositionCard[] =>
-    d.proposalPositions.map((p) => {
-      const position = {
-        ...p,
-        token: p.proposal.token,
-        pool: p.proposal.basicPool,
-        exchanges: p.proposal.exchanges.reduce((acc, e) => {
-          if (e.exchanges && e.exchanges.length > 0) {
-            return [...acc, ...e.exchanges]
-          }
-          return acc
-        }, []),
-      }
-      delete position.proposal
-
-      return position
-    })
-
   const [{ data, loading }, fetchMore] = useQueryPagination(
     RiskyPositionsQuery,
     variables,
-    prepareNewData
+    prepareRiskyPositions
   )
 
   const loader = useRef<any>()
